ps: Refactor mappings to `mapping.rs`
